4.21

+++ part4/blog/controllers/blogs.js
@@ -31,8 +31,23 @@ response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+    if(!blog) {
+      return response.status(404).json({ error: 'blog not found' })
+    }
+    
+    if(blog.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndDelete(request.params.id)
+      return response.status(204).end()
+    }
+
+    response.status(401).json({ error: 'unauthorized' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
